enhance git actions

+++ scripts/process.sh
@@ -2,8 +2,10 @@
 # Scan data for ymal files and combine them then transform to csv and json
 # ./process.sh ../data
 
-./libs/create-data.sh "$1" json > "${1}/index.json"
-./libs/create-data.sh "$1" csv > "${1}/index.csv"
+$(dirname $0)/libs/create-data.sh "$1" json > "${1}/index.json"
+$(dirname $0)/libs/create-data.sh "$1" csv > "${1}/index.csv"
 
-rm "${1}/index.db"
+if [ -f "${1}/index.db" ]; then
+  rm "${1}/index.db"
+fi
 sqlite-utils insert "${1}/index.db" data "${1}/index.json"
